ref: add metrics prometheus exporter
